Move MultiWriteSyncer alongside the rest of WriteSyncer
